updating with changes from code review and updated docs

+++ lib/inc/staticAllocator.hpp
@@ -4,9 +4,23 @@
 #include <cassert>
 #include <cstddef>
 
+/**
+ * 
+ * An alloctor which uses static storage for storing the elements.
+ * 
+ * @tparam T The type of the elements to be stored.
+ * @tparam size The amount of elements that can be stored.
+ * 
+ */
 template <typename T, size_t size>
 class StaticAllocator {
 public:
+    /**
+     * Returns a pointer to an empty T.
+     * 
+     * @return A pointer to a T if space is available. If the buffer is full a nullptr will be returned.
+     * 
+     */
     [[nodiscard]] T* get()
     {
         Element* element;
@@ -25,6 +39,12 @@ public:
         return &element->data;
     }
 
+    /**
+     * Frees an element in the buffer.
+     * 
+     * @param element The element to be freed.
+     * 
+     */
     void free(T* element)
     {
         auto index = std::find_if(data.begin(), data.end(), [element](Element& ele) { return &ele == reinterpret_cast<StaticAllocator<T, size>::Element*>(element); });
@@ -32,7 +52,10 @@ public:
         index->inUse = false;
     }
 
-    uint32_t countFreeElements()
+    /**
+     * Returns the amount of elements in the buffer which are still free. 
+     */
+    size_t countFreeElements()
     {
         return std::count_if(data.begin(), data.end(), [](const Element& ele) {
             return !ele.inUse;
